requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
Brotli==1.0.9
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
distlib==0.3.1
Django==3.2a1
django-background-tasks==1.2.5
django-bootstrap-datepicker==1.4.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3-datetimepicker==2.2.3
django-bootstrap4==2.3.1
django-common-helpers==0.9.2
django-compat==1.0.15
django-crispy-forms==1.11.0
django-cron==0.5.1
django-crontab==0.7.1
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
django-searchable-select==1.5.0
django-session-timeout==0.1.0
django-sitetree==1.16.0
django-slick-reporting==0.5.7
django-tables==0.2
django-tables2==2.3.4
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
filelock==3.0.12
form==0.0.1
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.20.0
pandas==1.2.2
plotly==4.14.3
prompt-toolkit==3.0.17
protobuf==3.15.1
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
simplejson==3.17.2
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
vine==5.0.0
virtualenv==20.4.2
wcwidth==0.2.5
Werkzeug==1.0.1